[FE]feat: typescript
