--- client/src/pages/Login.js
@@ -1,4 +1,0 @@
-function Login() {
-  return <>로그인</>;
-}
-export default Login;
