+++ client/src/pages/Login.tsx
@@ -1,0 +1,6 @@
+import Layout from 'components/Layout';
+
+function Login() {
+  return <Layout>로그인</Layout>;
+}
+export default Login;
