+++ client/src/pages/NoticeDetail.tsx
